tests/requirements.txt
robotframework>=5,<6
robotframework-robocop
python-dotenv